Crate sunrise_libuser

Userspace library

Provides an allocator, various lang items.

Re-exports

pub use futures as futures_rs;

Modules

ahci
Auto-generated documentation

allocator
Heap allocator.

argv
Argument handling

caps
Kernel Capabilities declaration

error
Error handling

example
Auto-generated documentation

fs
Auto-generated documentation

futures
Futures Executor

io
The IO interface

ipc
Core IPC Routines

keyboard
Auto-generated documentation

ldr
Auto-generated documentation

log_impl
Implementation for the log crate

loop_future
Definition of the LoopFn combinator, implementing Future loops.

mem
Memory

ps2
PS2 Keyboard APIs

sm
Auto-generated documentation

syscalls
Syscall Wrappers

terminal
Terminal rendering APIs

thread_local_storage
Thread Local Storage on x86

threads
Low-level api to create threads and start them.

time
Auto-generated documentation

twili
Auto-generated documentation

types
Core kernel types.

vi
Auto-generated documentation

window
Window creation and drawing APIs

zero_box
A module that allocates zeroed types on the heap without copying them from the stack first.

Macros

capabilities
Define the capabilities array in the .kernel_caps section. Has the following syntax:

kip_header
Define the kernel built-ins in the .kip_header section. Has the following syntax:

Functions

real_start⚠
calls logger initialization, main, and finally exits the process.