Module sunrise_libuser::example

Auto-generated documentation

Structs

IExample1Proxy
Here's a simple example of an interface. This is a service that can be connected to by connecting to "hello:1" with sm's get_service function.

IExample2Proxy
Let's make things a bit more interesting by adding a function or two.

IExample3SubsessionProxy
An instance of this interface cannot be acquired by asking SM. Instead, the user has to go through IExample3's function in order to get one.

IExample3Proxy
Interfaces can also return other interfaces. We call an instance of an interface an object. This is often used to give each interface its own state. For instance, a filesystem may return an object of interface file when the user requests to open one.

IExample4Proxy
There are two more kinds of types that are important: buffers and handles.

Traits

IExample1
Here's a simple example of an interface. This is a service that can be connected to by connecting to "hello:1" with sm's get_service function.

IExample1Async
Here's a simple example of an interface. This is a service that can be connected to by connecting to "hello:1" with sm's get_service function.

IExample2
Let's make things a bit more interesting by adding a function or two.

IExample2Async
Let's make things a bit more interesting by adding a function or two.

IExample3Subsession
An instance of this interface cannot be acquired by asking SM. Instead, the user has to go through IExample3's function in order to get one.

IExample3SubsessionAsync
An instance of this interface cannot be acquired by asking SM. Instead, the user has to go through IExample3's function in order to get one.

IExample3
Interfaces can also return other interfaces. We call an instance of an interface an object. This is often used to give each interface its own state. For instance, a filesystem may return an object of interface file when the user requests to open one.

IExample3Async
Interfaces can also return other interfaces. We call an instance of an interface an object. This is often used to give each interface its own state. For instance, a filesystem may return an object of interface file when the user requests to open one.

IExample4
There are two more kinds of types that are important: buffers and handles.

IExample4Async
There are two more kinds of types that are important: buffers and handles.