Module sunrise_libuser::fs

Auto-generated documentation

Structs

DirectoryEntry
Represent an entry inside a directory.

DirectoryEntryType
Represent the type of a given resource when walking a directory.

FileSystemType
Represent the type of a filesystem.

FileTimeStampRaw
Represent the attached timestamps on a given resource.

IDirectoryProxy
Represent a directory in a filesystem.

IFileProxy
Represent a file in a filesystem.

IFileSystemProxy
Represent a filesystem.

IFileSystemServiceProxy
Entry point of the file system interface.

IStorageProxy
This is the interface for a raw device, usually a block device.

Traits

IDirectory
Represent a directory in a filesystem.

IDirectoryAsync
Represent a directory in a filesystem.

IFile
Represent a file in a filesystem.

IFileAsync
Represent a file in a filesystem.

IFileSystem
Represent a filesystem.

IFileSystemAsync
Represent a filesystem.

IFileSystemService
Entry point of the file system interface.

IFileSystemServiceAsync
Entry point of the file system interface.

IStorage
This is the interface for a raw device, usually a block device.

IStorageAsync
This is the interface for a raw device, usually a block device.

Type Definitions

DiskId
A disk unique identifier.

FileSystemPath
Represent a path in a filesystem.

PartitionId
A partition unique identifier on a disk.