Struct sunrise_libuser::futures::WorkQueue

pub struct WorkQueue<'a>(Arc<Mutex<VecDeque<WorkItem<'a>>>>);

A WorkQueue represents a handle to a WaitableManager on which you can spawn new Futures with WorkQueue::spawn() or put the current future to sleep until a handle is signaled through HandleRef::wait_async().

This handle may be cloned - it will still point to the same WaitableManager. It may be shared with other threads, sent to other event loops, etc... in order to implement message passing.

Internally, a WorkQueue is an (Arc'd) deque of WorkItem, which the event loop will pop from in order to drive the scheduler.

Implementations

impl<'a> WorkQueue<'a>

pub fn spawn(&self, future: FutureObj<'a, ()>)

Spawn a top-level future on the event loop. The future will be polled once on spawn. Once the future is spawned, it will be owned by the WaitableManager.

pub(crate) fn simple(self) -> SimpleWorkQueue

Turn this WorkQueue into a SimpleWorkQueue.