Function sunrise_libuser::ipc::server::managed_port_handler

pub fn managed_port_handler<T, DISPATCH>(
    work_queue: WorkQueue<'static>, 
    server_name: &str, 
    dispatch: DISPATCH
) -> Result<impl Future<Output = ()>, Error> where
    DISPATCH: for<'b> FutureCallback<(&'b mut T, WorkQueue<'static>, u32, &'b mut [u8]), Result<(), Error>>,
    DISPATCH: Clone + Unpin + Send + 'static,
    T: Default + Clone + Unpin + Send + 'static, 

Creates a port through syscalls::manage_named_port() with the given name, and returns a future which will handle the port - that is, it will continuously accept new sessions on the port, and create backing objects through T::default(), and spawn a top-level future handling that sesion with new_session_wrapper().