Function sunrise_libuser::ipc::server::new_session_wrapper

pub fn new_session_wrapper<T, DISPATCH>(
    work_queue: WorkQueue<'static>, 
    handle: ServerSession, 
    object: T, 
    dispatch: DISPATCH
) -> impl Future<Output = ()> + Send where
    DISPATCH: for<'b> FutureCallback<(&'b mut T, WorkQueue<'static>, u32, &'b mut [u8]), Result<(), Error>>,
    DISPATCH: Unpin + Send + Clone + 'static,
    T: Unpin + Send + Clone + 'static, 

Creates a new top-level future that handles session.

The returned future will continuously accept new incoming requests on the handle, call the dispatch function with the given object, and the request' cmdid and buffer, and finally reply to the request.

It may be used to open subsessions.