Module sunrise_libuser::mem

Memory

Low-level helpers to assist memory mapping, MMIOs and DMAs.

Constants

PAGE_SIZE
The size of page. Used to interface with the kernel.

Functions

find_free_address
Finds a free memory zone of the given size and alignment in the current process's virtual address space. Note that the address space is not reserved, a call to map_memory to that address space might fail if another thread maps to it first. It is recommended to use this function and the map syscall in a loop.

map_mmio
Maps a Mmio struct in the virtual memory of this process.

virt_to_phys
Gets the physical address of a structure from its virtual address, preserving offset in the page.