Struct sunrise_libuser::threads::ThreadContext

#[repr(C)]pub struct ThreadContext {
    entry_point: fn(_: usize),
    arg: usize,
    stack: Option<StackContext>,
    tls_elf: Once<TlsElf>,
    thread_handle: Once<ThreadHandle>,
}

Structure holding the thread local context of a thread. Allocated at thread creation by the creator of the thread.

Fields
entry_point: fn(_: usize)

Pointer to the function this thread should execute after all its set-up in thread_trampoline is done.

arg: usize

The argument to call it with.

stack: Option<StackContext>

The stack used by this thread.

None for the main thread's stack, since it was not allocated by us and will never be freed as it'll be the last thread alive.

Some for every other thread.

tls_elf: Once<TlsElf>

The thread local storage of this thread.

This is where #[thread_local] statics live.

thread_handle: Once<ThreadHandle>

The ThreadHandle of this thread.

The thread needs to be able to access its own ThreadHandle at anytime to be able to use mutexes.