Struct sunrise_libuser::time::TimeZoneServiceProxy

pub struct TimeZoneServiceProxy(ClientSession);

TimeZone service object

Implementations

impl TimeZoneServiceProxy

pub fn clone_current_object(&self) -> Result<Self, Error>

Clones the current object, returning a new handle. The returned handle has its own IPC buffer - it may be used concurrently with the original.

pub fn get_device_location_name(&self) -> Result<LocationName, Error>

Get the time zone name used on this devie

pub fn set_device_location_name(
&self,
unknown_0: LocationName
) -> Result<(), Error>

Set the time zone name used on this devie

pub fn get_total_location_name_count(&self) -> Result<u32, Error>

Get the total count of location name available

pub fn load_location_name_list(
&self,
unknown_0: u32,
unknown_1: &mut [LocationName]
) -> Result<u32, Error>

Load the list of location name available

pub fn load_timezone_rule(
&self,
unknown_0: LocationName,
unknown_1: &mut TimeZoneRule
) -> Result<(), Error>

Load a time zone rule

pub fn to_calendar_time(
&self,
unknown_0: PosixTime,
unknown_1: &TimeZoneRule
) -> Result<(CalendarTime, CalendarAdditionalInfo), Error>

Convert a PosixTime to a local time using a TimeZoneRule

pub fn to_calendar_time_with_my_rule(
&self,
unknown_0: PosixTime
) -> Result<(CalendarTime, CalendarAdditionalInfo), Error>

Convert a PosixTime to a local time using the device TimeZoneRule

pub fn to_posix_time(
&self,
unknown_0: CalendarTime,
unknown_1: &TimeZoneRule
) -> Result<PosixTime, Error>

Convert a local time to a PosixTime using a TimeZoneRule.

pub fn to_posix_time_with_my_rule(
&self,
unknown_0: CalendarTime
) -> Result<PosixTime, Error>

Convert a local time to a PosixTime using the device TimeZoneRule