Trait sunrise_libuser::time::TimeZoneServiceAsync

pub trait TimeZoneServiceAsync {
    fn get_device_location_name<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>
    ) -> FutureObj<'a, Result<LocationName, Error>>;

    fn set_device_location_name<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>, 
        unknown_0: LocationName
    ) -> FutureObj<'a, Result<(), Error>>;

    fn get_total_location_name_count<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>
    ) -> FutureObj<'a, Result<u32, Error>>;

    fn load_location_name_list<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>, 
        unknown_0: u32, 
        unknown_1: &'a mut [LocationName]
    ) -> FutureObj<'a, Result<u32, Error>>;

    fn load_timezone_rule<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>, 
        unknown_0: LocationName, 
        unknown_1: &'a mut TimeZoneRule
    ) -> FutureObj<'a, Result<(), Error>>;

    fn to_calendar_time<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>, 
        unknown_0: PosixTime, 
        unknown_1: &'a TimeZoneRule
    ) -> FutureObj<'a, Result<(CalendarTime, CalendarAdditionalInfo), Error>>;

    fn to_calendar_time_with_my_rule<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>, 
        unknown_0: PosixTime
    ) -> FutureObj<'a, Result<(CalendarTime, CalendarAdditionalInfo), Error>>;

    fn to_posix_time<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>, 
        unknown_0: CalendarTime, 
        unknown_1: &'a TimeZoneRule
    ) -> FutureObj<'a, Result<PosixTime, Error>>;

    fn to_posix_time_with_my_rule<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>, 
        unknown_0: CalendarTime
    ) -> FutureObj<'a, Result<PosixTime, Error>>;

    fn dispatch<'a>(
        &'a mut self, 
        work_queue: WorkQueue<'static>, 
        cmdid: u32, 
        buf: &'a mut [u8]
    ) -> FutureObj<Result<(), Error>> { ... }
}

TimeZone service object

Required methods

fn get_device_location_name<'a>(
&'a mut self,
work_queue: WorkQueue<'static>
) -> FutureObj<'a, Result<LocationName, Error>>

Get the time zone name used on this devie

fn set_device_location_name<'a>(
&'a mut self,
work_queue: WorkQueue<'static>,
unknown_0: LocationName
) -> FutureObj<'a, Result<(), Error>>

Set the time zone name used on this devie

fn get_total_location_name_count<'a>(
&'a mut self,
work_queue: WorkQueue<'static>
) -> FutureObj<'a, Result<u32, Error>>

Get the total count of location name available

fn load_location_name_list<'a>(
&'a mut self,
work_queue: WorkQueue<'static>,
unknown_0: u32,
unknown_1: &'a mut [LocationName]
) -> FutureObj<'a, Result<u32, Error>>

Load the list of location name available

fn load_timezone_rule<'a>(
&'a mut self,
work_queue: WorkQueue<'static>,
unknown_0: LocationName,
unknown_1: &'a mut TimeZoneRule
) -> FutureObj<'a, Result<(), Error>>

Load a time zone rule

fn to_calendar_time<'a>(
&'a mut self,
work_queue: WorkQueue<'static>,
unknown_0: PosixTime,
unknown_1: &'a TimeZoneRule
) -> FutureObj<'a, Result<(CalendarTime, CalendarAdditionalInfo), Error>>

Convert a PosixTime to a local time using a TimeZoneRule

fn to_calendar_time_with_my_rule<'a>(
&'a mut self,
work_queue: WorkQueue<'static>,
unknown_0: PosixTime
) -> FutureObj<'a, Result<(CalendarTime, CalendarAdditionalInfo), Error>>

Convert a PosixTime to a local time using the device TimeZoneRule

fn to_posix_time<'a>(
&'a mut self,
work_queue: WorkQueue<'static>,
unknown_0: CalendarTime,
unknown_1: &'a TimeZoneRule
) -> FutureObj<'a, Result<PosixTime, Error>>

Convert a local time to a PosixTime using a TimeZoneRule.

fn to_posix_time_with_my_rule<'a>(
&'a mut self,
work_queue: WorkQueue<'static>,
unknown_0: CalendarTime
) -> FutureObj<'a, Result<PosixTime, Error>>

Convert a local time to a PosixTime using the device TimeZoneRule

Loading content...
Provided methods

fn dispatch<'a>(
&'a mut self,
work_queue: WorkQueue<'static>,
cmdid: u32,
buf: &'a mut [u8]
) -> FutureObj<Result<(), Error>>

Handle an incoming IPC request.

Loading content...