Struct sunrise_libuser::twili::ITwiliServiceProxy

pub struct ITwiliServiceProxy(ClientSession);

The Twili Service is responsible for providing the stdin/stdout/stderr pipes to the various subprocesses. It is freely inspired by the twili switch sysmodule.

Implementations

impl ITwiliServiceProxy

pub fn raw_new() -> Result<ITwiliServiceProxy, Error>

Creates a new ITwiliServiceProxy by connecting to the twili service.

pub fn new() -> Result<&'static ITwiliServiceProxy, Error>

Acquires the shared handle to the twili service - connecting if it wasn't already.

impl ITwiliServiceProxy

pub fn clone_current_object(&self) -> Result<Self, Error>

Clones the current object, returning a new handle. The returned handle has its own IPC buffer - it may be used concurrently with the original.

pub fn open_pipes(&self) -> Result<(IPipeProxy, IPipeProxy, IPipeProxy), Error>

Recover the stdin/stdout/stderr pipes for the current process.

If none were registered, returns an error.

pub fn create_pipe(&self) -> Result<(IPipeProxy, IPipeProxy), Error>

Creates a pipe whose write gets buffered and returned to the read side.

Tryting to write to the read side or read from the write side will return an error.