Struct sunrise_libuser::types::ClientSession

#[repr(transparent)]
pub struct ClientSession(pub Handle);

The client side of an IPC session.

Usually obtained by connecting to a service through the sm: service manager. However, an anonymous session pair might be created through the create_session syscall, or by calling connect_to_named_port.

Methods

impl ClientSession

pub fn send_sync_request_with_user_buffer(
&self,
buf: &mut [u8]
) -> Result<(), Error>

Send an IPC request to the handle, and wait for a response. The passed buffer should contain the request on input, and will contain the reply on output.

This is a low-level primitives that is usually wrapped by a higher-level library. Look at the ipc module for more information on the IPC message format.

pub fn into_handle(self) -> Handle

Consumes the session, returning the underlying handle. Note that closing a Handle without sending a close IPC message will leak the object in the sysmodule. You should always reconstruct the ClientSession from the Handle before dropping it.

pub fn try_clone(&self) -> Result<ClientSession, Error>

Clones the current object, returning a new handle. The returned handle has its own IPC buffer - it may be used concurrently with the original.