Struct sunrise_libuser::types::ServerSession

#[repr(transparent)]pub struct ServerSession(pub Handle);

The server side of an IPC session.

Usually obtained by calling accept, but may also be obtained by calling the create_session syscall, providing a server/client session pair.

Implementations

impl ServerSession

pub fn receive(
&self,
buf: &mut [u8],
timeout: Option<usize>
) -> Result<(), Error>

Receives an IPC request from the session, waiting if none are available yet. The buffer should contain an empty message, optionally containing a C descriptor, and will contain the reply on output.

If a C descriptor is provided, it will be used as the buffer to copy the request's X descriptor into.

This is a low-level primitives that is usually wrapped by a higher-level library. Look at the ipc module for more information on the IPC message format.

pub fn reply(&self, buf: &mut [u8]) -> Result<(), Error>

Replies to an IPC request on the given session. If the given session did not have a pending request, this function will error out.

This is a low-level primitives that is usually wrapped by a higher-level library. Look at the ipc module for more information on the IPC message format.

pub fn wait_async(
&self,
queue: WorkQueue
) -> impl Future<Output = Result<(), Error>> + Unpin

Waits for the server to receive a request.

Once this function returns, calling ServerSession::receive() is guaranteed not to block.

Panics

Panics if used from outside the context of a Future spawned on a libuser future executor. Please make sure you only call this function from a future spawned on a WaitableManager.